Update system package manager commands for Linux and macOS

+++ .zshrc
@@ -42,10 +42,16 @@ dotfiles() {
 }
 
 update() {
+  if [ "$(uname)" == "Linux" ]; then
     sudo apt update
     sudo apt upgrade -y
     sudo do-release-upgrade
     sudo apt autoremove -y
+  elif [ "$(uname)" == "Darwin" ]; then
+    brew update
+    brew upgrade
+    brew cleanup
+  fi
 }
 
 # Tools
